refactor(mga): split classifier_service utility grab-bag into cohesive modules
classifier_service.py had become a "utility service" — a leftover of the PR R2
split that held four unrelated concerns plus a re-export shim, with every helper
private (_-prefixed) yet imported across six sibling modules + the test suite
(private-by-convention used as a cross-module public API).

Split into cohesive modules with PUBLIC names:
- prompts.py          — GAME_VISUAL_CUES, GAME_FIRST_RULE, build_reference_text
- response_parsing.py — strip_json_fences, validate_aim_coord, validate_grid_index
- scope_guards.py     — check_game_map_consistency
- screenshots.py      — fetch_screenshot_bytes

All importers repointed: grid / single_image / aim / stand / throw-timing +
throw-technique consume the utilities; ingestion_orchestrator / lineup_service /
throw_localizer / technique_extractor import the classify_* entrypoints directly
from their own modules. classifier_service.py deleted — no back-compat shim.

Pure move + rename, no logic changes. ingestion_orchestrator.py held net-zero
(733 LOC, under the no-growth guard). Full backend suite: 776 passed.
